Use doubles instead of BigDecimals for numeric obs values (#120)

+++ core/api/src/test/java/org/n52/sos/util/builder/QuantityObservationValueBuilder.java
@@ -28,13 +28,11 @@
  */
 package org.n52.sos.util.builder;
 
-import java.math.BigDecimal;
-
 import org.joda.time.DateTime;
 import org.n52.sos.ogc.gml.time.TimeInstant;
 import org.n52.sos.ogc.om.SingleObservationValue;
-import org.n52.sos.ogc.om.values.Value;
 import org.n52.sos.ogc.om.values.QuantityValue;
+import org.n52.sos.ogc.om.values.Value;
 
 /**
  * @author <a href="mailto:e.h.juerrens@52north.org">Eike Hinderk
@@ -47,7 +45,7 @@ public class QuantityObservationValueBuilder {
         return new QuantityObservationValueBuilder();
     }
 
-    private Value<BigDecimal> quantity;
+    private Value<Double> quantity;
 
     private long phenomenonTime;
 
@@ -61,8 +59,8 @@ public class QuantityObservationValueBuilder {
         return this;
     }
 
-    public SingleObservationValue<BigDecimal> build() {
-        SingleObservationValue<BigDecimal> value = new SingleObservationValue<BigDecimal>();
+    public SingleObservationValue<Double> build() {
+        SingleObservationValue<Double> value = new SingleObservationValue<Double>();
         value.setValue(quantity);
         value.setPhenomenonTime(new TimeInstant(new DateTime(phenomenonTime)));
         return value;
